#25 refactoring main page
#25 refactoring main page
