+++ lib/app/dogdom_app.dart
@@ -1,5 +1,5 @@
 import 'package:dogdom/features/login/presentation/ui/login_screen.dart';
-import 'package:dogdom/features/main/presentation/ui/main_screen.dart';
+import 'package:dogdom/features/main/presentation/ui/main_page.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 
@@ -20,7 +20,7 @@ class DogdomApp extends StatelessWidget {
       initialRoute: LoginScreen.id,
       routes: {
         LoginScreen.id: (context) => LoginScreen(),
-        MainScreen.id: (context) => MainScreen(),
+        MainPage.id: (context) => MainPage(),
       },
     );
   }
