--- lib/features/main/presentation/bloc/main_event.dart
@@ -1,8 +1,0 @@
-import 'package:freezed_annotation/freezed_annotation.dart';
-
-part 'main_event.freezed.dart';
-
-@freezed
-class MainEvent with _$MainEvent {
-  factory MainEvent.init() = MainEventInit;
-}
