+++ lib/features/main/presentation/bloc/main_page_bloc.dart
@@ -1,14 +1,13 @@
 import 'package:dogdom/features/main/domain/interactors/main_interactor.dart';
-import 'package:dogdom/features/main/presentation/bloc/main_event.dart';
-import 'package:dogdom/features/main/presentation/bloc/main_state.dart';
+import 'package:dogdom/features/main/presentation/bloc/main_page_bloc_models.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:injectable/injectable.dart';
 
 @injectable
-class MainBloc extends Bloc<MainEvent, MainState> {
+class MainPageBloc extends Bloc<MainEvent, MainState> {
   final MainInteractor _interactor;
 
-  MainBloc(this._interactor) : super(MainState.empty()) {
+  MainPageBloc(this._interactor) : super(MainState.empty()) {
     on<MainEventInit>(_init);
   }
