+++ lib/features/main/presentation/bloc/main_page_bloc_models.dart
@@ -1,16 +1,17 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-part 'main_state.freezed.dart';
+part 'main_page_bloc_models.freezed.dart';
 
 @freezed
 class MainState with _$MainState {
-  MainStateData get data => this as MainStateData;
-
-  MainState._();
-
   factory MainState.empty() = MainStateEmpty;
 
   factory MainState.data({
     required String phone,
   }) = MainStateData;
+}
+
+@freezed
+class MainEvent with _$MainEvent {
+  factory MainEvent.init() = MainEventInit;
 }
