+++ lib/features/main/presentation/bloc/main_page_bloc_models.freezed.dart
@@ -3,7 +3,7 @@
 // ignore_for_file: type=lint
 // ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
 
-part of 'main_state.dart';
+part of 'main_page_bloc_models.dart';
 
 // **************************************************************************
 // FreezedGenerator
@@ -109,8 +109,8 @@ class _$MainStateEmptyCopyWithImpl<$Res> extends _$MainStateCopyWithImpl<$Res>
 
 /// @nodoc
 
-class _$MainStateEmpty extends MainStateEmpty {
-  _$MainStateEmpty() : super._();
+class _$MainStateEmpty implements MainStateEmpty {
+  _$MainStateEmpty();
 
   @override
   String toString() {
@@ -189,9 +189,8 @@ class _$MainStateEmpty extends MainStateEmpty {
   }
 }
 
-abstract class MainStateEmpty extends MainState {
+abstract class MainStateEmpty implements MainState {
   factory MainStateEmpty() = _$MainStateEmpty;
-  MainStateEmpty._() : super._();
 }
 
 /// @nodoc
@@ -227,8 +226,8 @@ class _$MainStateDataCopyWithImpl<$Res> extends _$MainStateCopyWithImpl<$Res>
 
 /// @nodoc
 
-class _$MainStateData extends MainStateData {
-  _$MainStateData({required this.phone}) : super._();
+class _$MainStateData implements MainStateData {
+  _$MainStateData({required this.phone});
 
   @override
   final String phone;
@@ -318,12 +317,172 @@ class _$MainStateData extends MainStateData {
   }
 }
 
-abstract class MainStateData extends MainState {
+abstract class MainStateData implements MainState {
   factory MainStateData({required String phone}) = _$MainStateData;
-  MainStateData._() : super._();
 
   String get phone;
   @JsonKey(ignore: true)
   $MainStateDataCopyWith<MainStateData> get copyWith =>
       throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+class _$MainEventTearOff {
+  const _$MainEventTearOff();
+
+  MainEventInit init() {
+    return MainEventInit();
+  }
+}
+
+/// @nodoc
+const $MainEvent = _$MainEventTearOff();
+
+/// @nodoc
+mixin _$MainEvent {
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function() init,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult Function()? init,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? init,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(MainEventInit value) init,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult Function(MainEventInit value)? init,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(MainEventInit value)? init,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $MainEventCopyWith<$Res> {
+  factory $MainEventCopyWith(MainEvent value, $Res Function(MainEvent) then) =
+      _$MainEventCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class _$MainEventCopyWithImpl<$Res> implements $MainEventCopyWith<$Res> {
+  _$MainEventCopyWithImpl(this._value, this._then);
+
+  final MainEvent _value;
+  // ignore: unused_field
+  final $Res Function(MainEvent) _then;
+}
+
+/// @nodoc
+abstract class $MainEventInitCopyWith<$Res> {
+  factory $MainEventInitCopyWith(
+          MainEventInit value, $Res Function(MainEventInit) then) =
+      _$MainEventInitCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class _$MainEventInitCopyWithImpl<$Res> extends _$MainEventCopyWithImpl<$Res>
+    implements $MainEventInitCopyWith<$Res> {
+  _$MainEventInitCopyWithImpl(
+      MainEventInit _value, $Res Function(MainEventInit) _then)
+      : super(_value, (v) => _then(v as MainEventInit));
+
+  @override
+  MainEventInit get _value => super._value as MainEventInit;
+}
+
+/// @nodoc
+
+class _$MainEventInit implements MainEventInit {
+  _$MainEventInit();
+
+  @override
+  String toString() {
+    return 'MainEvent.init()';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType && other is MainEventInit);
+  }
+
+  @override
+  int get hashCode => runtimeType.hashCode;
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function() init,
+  }) {
+    return init();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult Function()? init,
+  }) {
+    return init?.call();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? init,
+    required TResult orElse(),
+  }) {
+    if (init != null) {
+      return init();
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(MainEventInit value) init,
+  }) {
+    return init(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult Function(MainEventInit value)? init,
+  }) {
+    return init?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(MainEventInit value)? init,
+    required TResult orElse(),
+  }) {
+    if (init != null) {
+      return init(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class MainEventInit implements MainEvent {
+  factory MainEventInit() = _$MainEventInit;
 }
